# api/docs/swagger.yaml
definitions:
  client_service.Client:
    properties:
      additional_phone_number:
        type: string
      address:
        type: string
      created_at:
        type: string
      driving_license_number:
        type: string
      driving_number_expired:
        type: string
      driving_number_given_date:
        type: string
      driving_number_given_place:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_blocked:
        type: boolean
      last_name:
        type: string
      passport_number:
        type: string
      passport_pinfl:
        type: string
      phone_number:
        type: string
      photo:
        type: string
      propiska:
        type: string
      updated_at:
        type: string
    type: object
  client_service.CreateClient:
    properties:
      additional_phone_number:
        type: string
      address:
        type: string
      driving_license_number:
        type: string
      driving_number_expired:
        type: string
      driving_number_given_date:
        type: string
      driving_number_given_place:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      passport_number:
        type: string
      passport_pinfl:
        type: string
      phone_number:
        type: string
      photo:
        type: string
      propiska:
        type: string
    type: object
...
  client_service.GetListClientResponse:
    properties:
      clients:
        items:
          $ref: '#/definitions/client_service.Client'
        type: array
      count:
        type: integer
    type: object
  client_service.UpdateClient:
    properties:
      additional_phone_number:
        type: string
      address:
        type: string
      driving_license_number:
        type: string
      driving_number_expired:
        type: string
      driving_number_given_date:
        type: string
      driving_number_given_place:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      passport_number:
        type: string
      passport_pinfl:
        type: string
      phone_number:
        type: string
      photo:
        type: string
      propiska:
        type: string
    type: object
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
  models.UpdatePatch:
    properties:
      data:
        additionalProperties: true
        type: object
      id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /client:
    get:
      consumes:
      - application/json
      description: Get Client List
      operationId: get_client_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllClientResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/client_service.GetListClientResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Client List
      tags:
      - Client
  /client/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Client
      operationId: delete_client
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Client
      tags:
      - Client
    get:
      consumes:
      - application/json
      description: Get Client By ID
      operationId: get_client_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/client_service.Client'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Client By ID
      tags:
      - Client
    patch:
      consumes:
      - application/json
      description: Patch Client
      operationId: patch_client
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdatePatchRequestBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePatch'
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/client_service.Client'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Patch Client
      tags:
      - Client
    put:
      consumes:
      - application/json
      description: Update Client
      operationId: update_client
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateClientRequestBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/client_service.UpdateClient'
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/client_service.Client'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Client
      tags:
      - Client
...
  /user:
    post:
      consumes:
      - application/json
      description: Create User
      operationId: create_user
      parameters:
      - description: CreateClient
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/client_service.CreateClient'
      produces:
      - application/json
      responses:
        "200":
          description: GetUserBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/client_service.Client'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create User
      tags:
      - User
swagger: "2.0"